refactor: moving databricks to the new format
